Rename sei → seievm; add DB chain removal
Rename the 'sei' chain assets and references to 'seievm', update icon mappings, and centralize chain removal logic.

Key changes:
- Added SupportSQLiteDatabase.removeChain helper to delete all DB refs for a chain using GLOB JSON-safe patterns.
- Migration_85_86 now calls db.removeChain("sei").
- Updated Android IconUrlGeneration and tests to use seievm asset name.
- Renamed Android/iOS chain icon assets and updated Images/ChainImage mapping.
- Adjusted migration tests to assert new wallets_connections/in_app_notifications ids.

This ensures consistent handling of the Sei EVM chain and safer JSON matching during removals.
