+++ android/data/services/store/src/androidTest/kotlin/com/gemwallet/android/service/store/Migration_85_86Test.kt
@@ -46,10 +46,10 @@ class Migration_85_86Test {
         assertEquals(0, db.count("accounts", "chain = 'sei'"))
         assertEquals(1, db.count("accounts", "chain = 'seievm'"))
         assertEquals(1, db.count("wallets"))
-        assertEquals(0, db.count("wallets_connections", "instr(chains, '\"sei\"') > 0"))
         assertEquals(1, db.count("wallets_connections"))
-        assertEquals(0, db.count("in_app_notifications", "instr(item, '\"sei\"') > 0"))
+        assertEquals("seievm-connection", db.string("SELECT id FROM wallets_connections"))
         assertEquals(1, db.count("in_app_notifications"))
+        assertEquals("seievm-notification", db.string("SELECT id FROM in_app_notifications"))
         assertEquals(0, db.count("contacts_addresses", "chain = 'sei'"))
         assertEquals(1, db.count("contacts_addresses", "chain = 'seievm'"))
         assertEquals(1, db.count("transactions"))
