+++ android/gemcore/src/test/kotlin/com/gemwallet/android/ext/ChainExtTest.kt
@@ -16,7 +16,6 @@ class ChainExtTest {
         assertEquals(AssetType.ERC20, Chain.SeiEvm.assetType())
         assertEquals(EVMChain.SeiEvm, Chain.SeiEvm.toEVM())
         assertEquals(ChainType.Ethereum, Chain.SeiEvm.toChainType())
-        assertEquals("file:///android_asset/chains/icons/sei.svg", Chain.SeiEvm.getIconUrl())
     }
 
     @Test
