+++ ios/Packages/Style/Sources/Resources/Assets.xcassets/chains/seievm.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "sei.svg",
+      "filename" : "seievm.svg",
       "idiom" : "universal"
     }
   ],
